house_info_app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.5
googlemaps==4.10.0
Pillow==10.1.0
pydantic==2.5.0
